requirements.txt
Django>=1.8.0
djangorestframework>=3.1.0
